debug
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/apidb/OrthoMCLWorkflow/trunk@68903 735e2a04-f8fc-0310-8a1b-f2942603c481

+++ Main/lib/perl/WorkflowSteps/ExtractGroupPeripheralFastas.pm
@@ -4,6 +4,7 @@ package OrthoMCLWorkflow::Main::WorkflowSteps::ExtractGroupPeripheralFastas;
 
 use strict;
 use ApiCommonWorkflow::Main::WorkflowSteps::WorkflowStep;
+use File::Basename;
 
 
 sub run {
@@ -12,18 +13,22 @@ sub run {
     my $workflowDataDir = $self->getWorkflowDataDir();
 
     my $outputDir = $self->getParamValue('outputDir');
-    my $outputTar = $self->getParamValue('outputTar');
+    my $tarAndZip = $self->getBooleanParamValue('tarAndZip');
+    my $tarSize = $self->getConfig('proteinsPerTarFile');
 
     if ($undo) {
-	$self->runCmd($test, "rm -r $workflowDataDir/$outputDir");
-	$self->runCmd($test, "rm -r $workflowDataDir/$outputTar");
+	$self->runCmd($test, "rm -r $workflowDataDir/$outputDir") if -e "$workflowDataDir/$outputDir";
+	$self->runCmd($test, "rm -r $workflowDataDir/$outputDir.tar.gz") if $tarAndZip;
     } else{
       if ($test) {
-	$self->runCmd(0, "mkdir $workflowDataDir/$outputDir");
-	$self->runCmd(0, "touch $workflowDataDir/$outputTar");
+	$self->runCmd(0, "mkdir $workflowDataDir/$outputDir") unless $tarAndZip;
+	$self->runCmd(0, "touch $workflowDataDir/$outputDir.tar.gz") if $tarAndZip;
       } else {
-	$self->runCmd($test, "extractGroupFastaFiles --outputDir $workflowDataDir/$outputDir --peripheralsOnly");
-	$self->runCmd($test, "tar -zcf $workflowDataDir/${outputDir}.tar.gz $workflowDataDir/$outputDir");
-      }
+	$self->runCmd($test, "extractGroupFastaFiles --outputDir $workflowDataDir/$outputDir --peripheralsOnly --tarBall $tarSize"); #seqs per tarball, regardless of how many groups
+	my($baseDir, $path, $suffix) = fileparse("$workflowDataDir/$outputDir");
+	chdir "$path" || die "can't chdir to '$path'\n";
+	$self->runCmd($test, "tar -zcf $baseDir.tar.gz $baseDir") if $tarAndZip;
+ 	$self->runCmd($test, "rm -r $workflowDataDir/$outputDir") if ($tarAndZip && -e "$workflowDataDir/$outputDir");
+     }
     }
 }
